address review: require the quadrature C++ kernel, guard zero-cover trees, trim comments
- move the numpy quadrature implementation into the test suite as the differential oracle; the C extension is now required and a missing build raises an ImportError at construction (the implementation fallback is gone from the shipped explainer)
- LinearTreeSHAP refuses trees with unreachable (zero-cover) subtrees instead of silently returning NaN for some instances; QuadratureTreeSHAP ensembles must share one input precision
- delete the dense interventional kernel paths orphaned by the removed entry points, trim narrating comments across the tree package, and fix stale docstrings and test names
