+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Highlights of new Features
 
-- adds `QuadratureTreeSHAP` in `shapiq.tree.quadrature`: a quadrature-based path-dependent tree explainer (Wettenstein et al., 2026; TreeGrad, 2026) computing Shapley values and **any-order** Shapley interactions (`"SV"`, `"SII"`, `"k-SII"`, plus the Banzhaf indices `"BV"`/`"BII"` via point evaluation) that is numerically exact in float64 at any tree depth and any number of features per decision path — resolving the instability of the polynomial explainers [#545](https://github.com/mmschlk/shapiq/issues/545). Ships a C++ kernel (with a numpy reference implementation selectable via `implementation=`) that benchmarks faster than the existing explainers at every measured setting: ~2× vs `LinearTreeSHAP` for Shapley values and 100×+ vs `TreeSHAPIQ` for interactions, with order-4 interactions in milliseconds. Ensembles are computed in a single batched C++ kernel invocation over a shared feature space (no per-tree Python objects on the hot path), which benchmarks at parity with XGBoost's native `pred_contribs` TreeSHAP on a single thread. `QuadratureTreeSHAP` is the **default path-dependent algorithm** of `TreeExplainer`, computing all supported indices (`"SV"`, `"SII"`, `"k-SII"`, `"BV"`, `"BII"`) at any order natively — path-dependent Banzhaf indices no longer require the optional woodelf dependency, and the order- and size-based routing rules between the previous path-dependent implementations were removed. Also directly available via `shapiq.tree.QuadratureTreeSHAP`
+- adds `QuadratureTreeSHAP` in `shapiq.tree.quadrature`: a quadrature-based path-dependent tree explainer (Wettenstein et al., 2026; TreeGrad, 2026) computing Shapley values and **any-order** Shapley interactions (`"SV"`, `"SII"`, `"k-SII"`, plus the Banzhaf indices `"BV"`/`"BII"` via point evaluation) that is numerically exact in float64 at any tree depth and any number of features per decision path — resolving the instability of the polynomial explainers [#545](https://github.com/mmschlk/shapiq/issues/545). Ships a C++ kernel (required — a missing extension raises an `ImportError` at construction; the numpy reference implementation lives in the test suite as the differential oracle) that benchmarks faster than the existing explainers at every measured setting: ~2× vs `LinearTreeSHAP` for Shapley values and 100×+ vs `TreeSHAPIQ` for interactions, with order-4 interactions in milliseconds. Ensembles are computed in a single batched C++ kernel invocation over a shared feature space (no per-tree Python objects on the hot path), which benchmarks at parity with XGBoost's native `pred_contribs` TreeSHAP on a single thread. `QuadratureTreeSHAP` is the **default path-dependent algorithm** of `TreeExplainer`, computing all supported indices (`"SV"`, `"SII"`, `"k-SII"`, `"BV"`, `"BII"`) at any order natively — path-dependent Banzhaf indices no longer require the optional woodelf dependency, and the order- and size-based routing rules between the previous path-dependent implementations were removed. Also directly available via `shapiq.tree.QuadratureTreeSHAP`
 - harmonizes the tree computation algorithms: `LinearTreeSHAP`, `TreeSHAPIQ`, `QuadratureTreeSHAP`, and `InterventionalTreeSHAPIQ` now all validate their own models, accept ensembles as well as single trees (aggregating internally), take a `class_index` argument, and expose a common `explain(x)` entry point — each algorithm can be used independently of `TreeExplainer`, which now simply instantiates one explainer per mode. Their modules are named `computer` (`shapiq.tree.linear.computer`, `shapiq.tree.quadrature.computer`, `shapiq.tree.interventional.computer`), keeping `shapiq.tree.explainer` unambiguous as the user-facing `TreeExplainer` entry point; the public class imports (`from shapiq.tree import ...`) are unchanged
 - the interventional TreeSHAP-IQ C kernels now compute in float64 (previously float32 inputs, thresholds, and accumulation on the dense path); the three unused kernel entry points (`compute_interactions`, `compute_interactions_batched`, `compute_interactions_sparse`) were removed
-- adds numerical precision guards to the polynomial tree explainers: `LinearTreeSHAP` and `TreeSHAPIQ` now emit a `TreeNumericalPrecisionWarning` from 24 distinct features per decision path (measurably degrading values) and raise a `TreeNumericalPrecisionError` from 30 (unreliable values), instead of silently returning wrong attributions or crashing with an unexplained `LinAlgError` [#545](https://github.com/mmschlk/shapiq/issues/545). `TreeSHAPIQ`'s interpolation grid and `LinearTreeSHAP`'s N-matrix construction are now sized by the maximum number of distinct features per path (instead of depth-derived bounds), which removes latent `LinAlgError` crashes for deep trees whose individual paths stay feature-narrow and slightly improves their accuracy
+- adds numerical precision guards to the polynomial tree explainers: `LinearTreeSHAP` and `TreeSHAPIQ` now emit a `TreeNumericalPrecisionWarning` from 24 distinct features per decision path (measurably degrading values) and raise a `TreeNumericalPrecisionError` from 30 (unreliable values), instead of silently returning wrong attributions or crashing with an unexplained `LinAlgError` [#545](https://github.com/mmschlk/shapiq/issues/545). `TreeSHAPIQ`'s interpolation grid and `LinearTreeSHAP`'s N-matrix construction are now sized by the maximum number of distinct features per path (instead of depth-derived bounds), which removes latent `LinAlgError` crashes for deep trees whose individual paths stay feature-narrow and slightly improves their accuracy. `LinearTreeSHAP` additionally refuses trees containing unreachable (zero-cover) subtrees, whose infinite edge weights its polynomial representation cannot express (previously it silently returned NaN attributions for some instances, e.g. on CatBoost models with unreachable oblivious-tree regions)
 - adds a Woodelf fast path to `shapiq.tree.TreeExplainer` (Nadel and Wettenstein, 2026; Wettenstein et al., 2026). Larger interventional inputs are computed by the vectorized Woodelf algorithms, and `TreeExplainer` now supports the Banzhaf indices `"BV"` and `"BII"`. `TreeExplainer.explain_X` explains all instances in one vectorized run and returns the new `InteractionValuesBatch`: a lazy sequence of one `InteractionValues` per instance whose `.values` attribute exposes the raw vectorized arrays (`Explainer.explain_X` is now typed as returning a `Sequence[InteractionValues]`). The fast path requires the new optional `tree` dependency group (`pip install shapiq[tree]` — `woodelf-explainer`); without it, the explainer falls back to the shapiq implementation with a `WoodelfNotAvailableWarning`. The new `backend` parameter (`"auto"`, `"woodelf"`, `"shapiq"`) overrides the automatic routing [#572](https://github.com/mmschlk/shapiq/pull/572)
 - adds the `ConfoundingSHAP` local and global explanation game in `shapiq_games.benchmark.causal_xai` for the first causal machine learning based value functions (Brockschmidt et al., 2026) based on the [paper](https://arxiv.org/abs/2605.10533).
 - adds the `LeverageSHAP` approximator in `shapiq.approximator.regression` for Shapley value estimation via leverage score sampling (Musco and Witter, 2025) [#524](https://github.com/mmschlk/shapiq/pull/524)
