+++ src/shapiq/tree/interventional/cext/interventional.cpp
@@ -44,76 +44,6 @@ namespace algorithms
         return num_features + (i * num_features - i * (i + 1) / 2) + (j - i - 1);
     }
 
-    void first_order_update(const StackFrame &frame, double value, double *interactions, inter_weights::WeightCache &weight_cache, int num_features, IndexType index, int max_order, int verbose)
-    {
-        const double wE = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 0, 1, index, max_order));
-        const double wR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 1, 1, index, max_order));
-
-        frame.E.for_each_set_bit([&](uint64_t feature)
-                                         { interactions[feature] += value * wE; });
-        frame.R.for_each_set_bit([&](uint64_t feature)
-                                         { interactions[feature] += value * wR; });
-    }
-
-    void second_order_update(const StackFrame &frame,
-                             uint64_t *e_buffer, uint64_t *r_buffer,
-                             uint64_t e_count, uint64_t r_count,
-                             double value, double *interactions, inter_weights::WeightCache &weight_cache, int num_features, IndexType index, int max_order, int verbose)
-    {
-
-        // Main Effect updates (diagonal elements)
-        const double wE = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 0, 1, index, max_order));
-        const double wR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 1, 1, index, max_order));
-        frame.E.for_each_set_bit([&](uint64_t feature)
-                                 {
-            int idx = get_interaction_index(static_cast<int>(feature), static_cast<int>(feature), num_features, max_order);
-            interactions[idx] += value * wE; });
-        frame.R.for_each_set_bit([&](uint64_t feature)
-                                 {
-            int idx = get_interaction_index(static_cast<int>(feature), static_cast<int>(feature), num_features, max_order);
-            interactions[idx] += value * wR; });
-
-        // Fill the buffers with the feature indices in E and R.
-        // This avoids repeated memory allocations during the interaction updates.
-        frame.E.fill_buffer(e_buffer);
-        frame.R.fill_buffer(r_buffer);
-
-        // Interaction in E (upper triangle)
-        const double wEE = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 2, 0, 2, index, max_order));
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = i + 1; j < e_count; j++)
-            {
-                int idx = get_interaction_index(static_cast<int>(e_buffer[i]), static_cast<int>(e_buffer[j]), num_features, max_order);
-                interactions[idx] += value * wEE;
-            }
-        }
-        // Interactions in R (upper triangle)
-        const double wRR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 2, 2, index, max_order));
-        for (size_t i = 0; i < r_count; i++)
-        {
-            for (size_t j = i + 1; j < r_count; j++)
-            {
-                int idx = get_interaction_index(static_cast<int>(r_buffer[i]), static_cast<int>(r_buffer[j]), num_features, max_order);
-                interactions[idx] += value * wRR;
-            }
-        }
-        // Cross interactions (E × R)
-        const double wER = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 1, 2, index, max_order));
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = 0; j < r_count; j++)
-            {
-                if (e_buffer[i] == r_buffer[j])
-                    continue; // Skip if the same feature is in both E and R, as this would correspond to a main effect, not an interaction.
-                int idx = get_interaction_index(static_cast<int>(e_buffer[i]), static_cast<int>(r_buffer[j]), num_features, max_order);
-                interactions[idx] += value * wER;
-            }
-        }
-    }
-
-    // Compact index for order-3 triples (i < j < k):
-    //   index3(i,j,k,n) = n + n*(n-1)/2 + i + j*(j-1)/2 + k*(k-1)*(k-2)/6
     inline int get_interaction_index3(int i, int j, int k, int num_features)
     {
         // Ensure i < j < k
@@ -122,128 +52,6 @@ namespace algorithms
         if (i > j) std::swap(i, j);
         int base = num_features + num_features * (num_features - 1) / 2;
         return base + i + j * (j - 1) / 2 + k * (k - 1) * (k - 2) / 6;
-    }
-
-    void third_order_update(const StackFrame &frame,
-                             uint64_t *e_buffer, uint64_t *r_buffer,
-                             uint64_t e_count, uint64_t r_count,
-                             double value, double *interactions,
-                             inter_weights::WeightCache &weight_cache,
-                             int num_features, IndexType index, int max_order, int verbose)
-    {
-        // Precompute all 8 weight types at the top of the function to avoid repeated cache lookups.
-        // Order-1 weights
-        const double wE  = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 0, 1, index, max_order));
-        const double wR  = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 1, 1, index, max_order));
-        // Order-2 weights
-        const double wEE = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 2, 0, 2, index, max_order));
-        const double wRR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 2, 2, index, max_order));
-        const double wER = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 1, 2, index, max_order));
-        // Order-3 weights
-        const double wEEE = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 3, 0, 3, index, max_order));
-        const double wRRR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 0, 3, 3, index, max_order));
-        const double wEER = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 2, 1, 3, index, max_order));
-        const double wERR = static_cast<double>(weight_cache.get_weight(num_features, frame.e, frame.r, 1, 2, 3, index, max_order));
-
-        // Fill buffers
-        frame.E.fill_buffer(e_buffer);
-        frame.R.fill_buffer(r_buffer);
-
-        // === Order-1 contributions ===
-        for (size_t i = 0; i < e_count; i++)
-        {
-            int idx = static_cast<int>(e_buffer[i]);
-            interactions[idx] += value * wE;
-        }
-        for (size_t i = 0; i < r_count; i++)
-        {
-            int idx = static_cast<int>(r_buffer[i]);
-            interactions[idx] += value * wR;
-        }
-
-        // === Order-2 contributions ===
-        // EE pairs
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = i + 1; j < e_count; j++)
-            {
-                int idx = get_interaction_index(static_cast<int>(e_buffer[i]), static_cast<int>(e_buffer[j]), num_features, 2);
-                interactions[idx] += value * wEE;
-            }
-        }
-        // RR pairs
-        for (size_t i = 0; i < r_count; i++)
-        {
-            for (size_t j = i + 1; j < r_count; j++)
-            {
-                int idx = get_interaction_index(static_cast<int>(r_buffer[i]), static_cast<int>(r_buffer[j]), num_features, 2);
-                interactions[idx] += value * wRR;
-            }
-        }
-        // ER cross pairs (skip if same feature appears in both E and R)
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = 0; j < r_count; j++)
-            {
-                if (e_buffer[i] == r_buffer[j]) continue;
-                int idx = get_interaction_index(static_cast<int>(e_buffer[i]), static_cast<int>(r_buffer[j]), num_features, 2);
-                interactions[idx] += value * wER;
-            }
-        }
-
-        // === Order-3 contributions ===
-        // EEE triples
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = i + 1; j < e_count; j++)
-            {
-                for (size_t k = j + 1; k < e_count; k++)
-                {
-                    int idx = get_interaction_index3(static_cast<int>(e_buffer[i]), static_cast<int>(e_buffer[j]), static_cast<int>(e_buffer[k]), num_features);
-                    interactions[idx] += value * wEEE;
-                }
-            }
-        }
-        // RRR triples
-        for (size_t i = 0; i < r_count; i++)
-        {
-            for (size_t j = i + 1; j < r_count; j++)
-            {
-                for (size_t k = j + 1; k < r_count; k++)
-                {
-                    int idx = get_interaction_index3(static_cast<int>(r_buffer[i]), static_cast<int>(r_buffer[j]), static_cast<int>(r_buffer[k]), num_features);
-                    interactions[idx] += value * wRRR;
-                }
-            }
-        }
-        // EER triples: 2 from E, 1 from R — skip if any two features are the same
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = i + 1; j < e_count; j++)
-            {
-                for (size_t k = 0; k < r_count; k++)
-                {
-                    uint64_t rk = r_buffer[k];
-                    if (e_buffer[i] == rk || e_buffer[j] == rk) continue;
-                    int idx = get_interaction_index3(static_cast<int>(e_buffer[i]), static_cast<int>(e_buffer[j]), static_cast<int>(rk), num_features);
-                    interactions[idx] += value * wEER;
-                }
-            }
-        }
-        // ERR triples: 1 from E, 2 from R — skip if any two features are the same
-        for (size_t i = 0; i < e_count; i++)
-        {
-            for (size_t j = 0; j < r_count; j++)
-            {
-                if (e_buffer[i] == r_buffer[j]) continue;
-                for (size_t k = j + 1; k < r_count; k++)
-                {
-                    if (e_buffer[i] == r_buffer[k]) continue;
-                    int idx = get_interaction_index3(static_cast<int>(e_buffer[i]), static_cast<int>(r_buffer[j]), static_cast<int>(r_buffer[k]), num_features);
-                    interactions[idx] += value * wERR;
-                }
-            }
-        }
     }
 
     void enumerate_r_subsets(const uint64_t *r_features,
@@ -348,120 +156,6 @@ namespace algorithms
                     subset,
                     contribution,
                     interactions);
-            }
-        }
-    }
-
-    void compute_interactions(Tree tree, double *interactions,
-                               inter_weights::WeightCache &weight_cache,
-                               double *reference_data,
-                               double *explain_data,
-                               int num_features,
-                               IndexType index,
-                               int max_order,
-                               int verbose)
-    {
-        /**
-         * This function computes the first-order interactions for a given decision tree and input data.
-         * The algorithm computes the index based on the algorithm of Zern. 2023 (Algorithm 1) "Interventional SHAP Values and Interaction Values for Piecewise Linear Regression Trees "
-         * The output is an array of interaction values for each feature.
-         * This algorithm here keep storage of [A, N\B] pairs on the stack, where A and B are the sets of features which are necessary to reach the current node based on the explain and reference point, respectively.
-         */
-
-        std::vector<StackFrame> stack;
-        BitSet empty_A(num_features);
-        BitSet empty_B(num_features);
-
-        // Create buffers for the feature indices in E and R to avoid repeated memory allocations during the interaction updates.
-        uint64_t bit_buffer_E[64];
-        uint64_t bit_buffer_R[64];
-        std::vector<uint64_t> vector_buffer_E;
-        std::vector<uint64_t> vector_buffer_R;
-
-        uint64_t *e_buffer;
-        uint64_t *r_buffer;
-        uint64_t e_count;
-        uint64_t r_count;
-
-        stack.reserve(1000);                                    // Reserve space for 1000 stack frames to avoid frequent reallocations. Adjust this number based on expected tree depth and branching factor.
-        stack.push_back(StackFrame(0, empty_A, empty_B, 0, 0)); // Start with the root node (node_id = 0) and empty sets A and B
-        while (!stack.empty())
-        {
-            // std::move is used to efficiently transfer only the pointer of the memory allocated for the StackFrame on the heap instead of copying the entire StackFrame, which can be expensive if it contains large data structures.
-            // After std::move, the current_frame variable takes ownership of the StackFrame object that was previously owned by the last element of the stack vector.
-            // This allows us to pop the last element from the stack without having to copy its contents, improving performance.
-            StackFrame current_frame = std::move(stack.back()); // Using std::move to avoid unnecessary copying of the StackFrame when popping from the stack.
-            stack.pop_back();
-            int64_t node_id = current_frame.node_id;
-            const BitSet &E = current_frame.E;
-            const BitSet &R = current_frame.R;
-            uint64_t e = current_frame.e;
-            uint64_t r = current_frame.r;
-            e_count = E.num_bits();
-            r_count = R.num_bits();
-
-            bool is_leaf = tree.is_leaf(node_id);
-
-            if (!is_leaf)
-            {
-                int64_t feature_id = tree.features[node_id];
-                int64_t child_explain_point = tree.goes_left(explain_data[feature_id], node_id) ? tree.children_left[node_id] : tree.children_right[node_id];
-                int64_t child_reference_point = tree.goes_left(reference_data[feature_id], node_id) ? tree.children_left[node_id] : tree.children_right[node_id];
-                // 1. Case: Both points go to the same child node
-                if (child_explain_point == child_reference_point)
-                {
-                    stack.push_back(StackFrame(child_explain_point, E, R, e, r));
-                }
-                else
-                {
-                    // 2. Case: Points go to different child nodes
-                    //  Add feature_id to E iff. i not contained in R
-                    if (!R.contains(feature_id))
-                    {
-                        BitSet next_E = E;
-                        bool added_to_E = next_E.add(feature_id);
-                        stack.push_back(StackFrame(child_explain_point, next_E, R, e + (added_to_E ? 1 : 0), r));
-                    }
-                    // Add feature_id to B iff. i not contained in A
-                    if (!E.contains(feature_id))
-                    {
-                        BitSet next_R = R;
-                        bool added_to_R = next_R.add(feature_id);
-                        stack.push_back(StackFrame(child_reference_point, E, next_R, e, r + (added_to_R ? 1 : 0)));
-                    }
-                }
-            }
-            else
-            {
-                // Get the correct buffer for E and R based on their sizes. For small sets (up to 64 features), we can use the pre-allocated bit buffers.
-                // For larger sets, we need to resize the vector buffers and use their data pointers.
-                if (e_count <= 64)
-                {
-                    e_buffer = bit_buffer_E;
-                }
-                else
-                {
-                    vector_buffer_E.resize(e_count);
-                    e_buffer = vector_buffer_E.data();
-                }
-                if (r_count <= 64)
-                {
-                    r_buffer = bit_buffer_R;
-                }
-                else
-                {
-                    vector_buffer_R.resize(r_count);
-                    r_buffer = vector_buffer_R.data();
-                }
-                double leaf_value = tree.leaf_predictions[node_id];
-                if (max_order == 1)
-                    algorithms::first_order_update(current_frame, leaf_value, interactions, weight_cache, num_features, index, max_order, verbose);
-                else if (max_order == 2)
-                    algorithms::second_order_update(current_frame, e_buffer, r_buffer, e_count, r_count, leaf_value, interactions, weight_cache, num_features, index, max_order, verbose);
-                else if (max_order == 3)
-                    algorithms::third_order_update(current_frame, e_buffer, r_buffer, e_count, r_count, leaf_value, interactions, weight_cache, num_features, index, max_order, verbose);
-                else
-                    throw std::invalid_argument("Unsupported max_order: " + std::to_string(max_order));
             }
         }
     }
